메뉴 골라서 장바구니에 담기

+++ src/Cart.java
@@ -3,9 +3,9 @@ import java.util.List;
 public class Cart {
 
     String name;
-    int price;
+    double price;
 
-    public Cart(String name, int price) {
+    public Cart(String name, double price) {
         this.name = name;
         this.price = price;
     }
